Refactor a load_resources_json method
